Created the bases for the alignment points
Created the bases for all of the points that spin around to align to the moon or earth. They also have pivots set ups so they will rotate as I want them to. This should be a good base for being able to create a 3D printed model that can actually move